RPC, cosmetic: move wallet-related RPCs to new rpcwallet.cpp module
