+++ src/makefile.linux-mingw
@@ -63,6 +63,7 @@ OBJS= \
     obj/rpcdump.o \
     obj/rpcnet.o \
     obj/rpcmining.o \
+    obj/rpcwallet.o \
     obj/rpcrawtransaction.o \
     obj/script.o \
     obj/sync.o \
