+++ src/makefile.mingw
@@ -59,6 +59,7 @@ OBJS= \
     obj/rpcdump.o \
     obj/rpcnet.o \
     obj/rpcmining.o \
+    obj/rpcwallet.o \
     obj/rpcrawtransaction.o \
     obj/script.o \
     obj/sync.o \
